Implemented methods and constructor
Implemented a parameter constructor that allows
the user to specify the name of the device they
want to control with the controller. To the
Init() method, added more PID errors that needed
to be initialized. Added code to the UpdateError()
method to update PID errors based on CTE. Added
code to TotalError() method that calculates the
error for the steering or throttle device.
TotalError() accounts for whether the user
specified they are using the PID to control a
steering or throttle device. UpdateSteering()
and UpdateThrottle() methods were implemented
to compute steering value and throttle value.
These two methods are referenced in TotalError().