Added explanation of PID coefficients
In the comments, I added explanations for PID speed
and steering on why I chose the coefficients values
to be small or large and how they affect the car
driving in the simulator. Implemented a PID controller
for controlling the car steering and throttle. I
have two objects: pid and pid_speed. One calculates
the steer_value and the other calculates the
throttle_value, then passes that data into msgJson
object that is later sent to the car in simulator.